+++ heroku/inline/token_obtainment.py
@@ -11,18 +11,18 @@
 # 🔑 https://www.gnu.org/licenses/agpl-3.0.html
 
 import asyncio
-import aiohttp
 import logging
-import re
 import os
 import random
+import re
 import typing
 
-from .. import utils
-from .. import main
+import aiohttp
+
+from .. import main, utils
 from .._internal import fw_protect
-from .types import InlineUnit
 from . import utils as inutils
+from .types import InlineUnit
 
 if typing.TYPE_CHECKING:
     from ..inline.core import InlineManager