+++ heroku/modules/heroku_security.py
@@ -312,7 +312,7 @@ class HerokuSecurityMod(loader.Module):
                 (
                     self.strings("permissions_list").format(
                         "\n".join(
-                            "<emoji document_id=4974307891025543730>▫️</emoji>"
+                            "<tg-emoji emoji-id=4974307891025543730>▫️</tg-emoji>"
                             " <b>{}</b> <code>{}</code> <b>{}</b>".format(
                                 self.strings(rule["rule_type"]),
                                 rule["rule"],
@@ -1204,7 +1204,7 @@ class HerokuSecurityMod(loader.Module):
                 self.strings("rules").format(
                     "\n".join(
                         [
-                            "<emoji document_id=6037355667365300960>👥</emoji> <b><a"
+                            "<tg-emoji emoji-id=6037355667365300960>👥</tg-emoji> <b><a"
                             " href='{}'>{}</a> {} {} {}</b> <code>{}</code>".format(
                                 rule["entity_url"],
                                 utils.escape_html(rule["entity_name"]),
@@ -1216,7 +1216,7 @@ class HerokuSecurityMod(loader.Module):
                             for rule in self._client.dispatcher.security.tsec_chat
                         ]
                         + [
-                            "<emoji document_id=6037122016849432064>👤</emoji> <b><a"
+                            "<tg-emoji emoji-id=6037122016849432064>👤</tg-emoji> <b><a"
                             " href='{}'>{}</a> {} {} {}</b> <code>{}</code>".format(
                                 rule["entity_url"],
                                 utils.escape_html(rule["entity_name"]),
@@ -1230,7 +1230,7 @@ class HerokuSecurityMod(loader.Module):
                         + [
                             "\n".join(
                                 [
-                                    "<emoji document_id=5870704313440932932>🔒</emoji>"
+                                    "<tg-emoji emoji-id=5870704313440932932>🔒</tg-emoji>"
                                     " <code>{}</code> <b>{} {} {}</b> <code>{}</code>".format(
                                         utils.escape_html(group.name),
                                         self._convert_time(