+++ heroku/utils/git.py
@@ -4,14 +4,16 @@
 # You can redistribute it and/or modify it under the terms of the GNU AGPLv3
 # 🔑 https://www.gnu.org/licenses/agpl-3.0.html
 
-import logging
-import typing
 import asyncio
-import subprocess
+import logging
 import os
+import subprocess
+import typing
 
 import git
 import herokutl
+
+from .. import version
 
 parser = herokutl.utils.sanitize_parse_mode("html")
 logger = logging.getLogger(__name__)
@@ -80,7 +82,7 @@ def get_git_status() -> str:
 
         if not output:
             return "Clean"
-        
+
         count = len(output.splitlines())
         word = "file" if count == 1 else "files"
         return f"{count} {word} modified"
@@ -115,3 +117,10 @@ def get_commit_count() -> int:
         return len(list(repo.iter_commits()))
     except Exception:
         return 0
+
+def is_up_to_date():
+    repo = git.Repo(search_parent_directories=True)
+    diff = any(
+        repo.iter_commits(f"HEAD..origin/{version.branch}", max_count=1)
+    )
+    return not diff