+++ heroku/utils/heroku.py
@@ -6,6 +6,7 @@
 
 import logging
 import os
+
 import herokutl
 
 parser = herokutl.utils.sanitize_parse_mode("html")