+++ CHANGELOG.md
@@ -78,6 +78,11 @@
 - backupall send backup in backupchat
 - fully remove save to fs
 - optimize username handling & fix logic
+- protection for session files
+- add premium emoji in some cases with inline
+- add developer in help
+- add folder for config
+- remove pyproject.toml
 
 ## 🪐 Heroku 1.7.2
 - added autoupdate